2020 day 16 (perl)
I haven't totally understood this problem yet, because it's a fucking
mess. Lines 28-42 feel like they're nested inside-out. 'Inverting' the
sense of the hash I just spent all that time constructing is gross. A
proper depth-first search is needed once the lists of candidates is
available. The key-value relationship in %used feels backward.